requirements.txt
crewai[tools]>=1.10.0
python-dotenv>=1.0.0
